added individual repr for seggregated clients
