+++ packages/syft/src/syft/client/client.py
@@ -24,7 +24,6 @@ from typing_extensions import Self
 # relative
 from .. import __version__
 from ..abstract_node import AbstractNode
-from ..img.base64 import base64read
 from ..node.credentials import SyftSigningKey
 from ..node.credentials import SyftVerifyKey
 from ..node.credentials import UserLoginCredentials
@@ -43,7 +42,6 @@ from ..service.user.user_service import UserService
 from ..types.grid_url import GridURL
 from ..types.syft_object import SYFT_OBJECT_VERSION_1
 from ..types.uid import UID
-from ..util.fonts import fonts_css
 from ..util.logger import debug
 from ..util.telemetry import instrument
 from ..util.util import thread_ident
@@ -580,100 +578,6 @@ class SyftClient:
             uid = proxy_target_uid
             return f"<{client_type} - <{uid}>: via {self.id} {self.connection}>"
         return f"<{client_type} - {self.name} <{uid}>: {self.connection}>"
-
-    def _repr_html_(self) -> str:
-        guest_commands = """
-        <li><span class='syft-code-block'>&lt;your_client&gt;.datasets</span> - list datasets</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.code</span> - list code</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.login</span> - list projects</li>
-        <li>
-            <span class='syft-code-block'>&lt;your_client&gt;.code.submit?</span> - display function signature
-        </li>"""
-        ds_commands = """
-        <li><span class='syft-code-block'>&lt;your_client&gt;.datasets</span> - list datasets</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.code</span> - list code</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.projects</span> - list projects</li>
-        <li>
-            <span class='syft-code-block'>&lt;your_client&gt;.code.submit?</span> - display function signature
-        </li>"""
-
-        do_commands = """
-        <li><span class='syft-code-block'>&lt;your_client&gt;.projects</span> - list projects</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.requests</span> - list requests</li>
-        <li><span class='syft-code-block'>&lt;your_client&gt;.users</span> - list users</li>
-        <li>
-            <span class='syft-code-block'>&lt;your_client&gt;.requests.submit?</span> - display function signature
-        </li>"""
-
-        # TODO: how to select ds/do commands based on self.__user_role
-
-        if (
-            self.user_role.value == ServiceRole.NONE.value
-            or self.user_role.value == ServiceRole.GUEST.value
-        ):
-            commands = guest_commands
-        elif (
-            self.user_role is not None
-            and self.user_role.value == ServiceRole.DATA_SCIENTIST.value
-        ):
-            commands = ds_commands
-        elif (
-            self.user_role is not None
-            and self.user_role.value >= ServiceRole.DATA_OWNER.value
-        ):
-            commands = do_commands
-
-        command_list = f"""
-        <ul style='padding-left: 1em;'>
-            {commands}
-        </ul>
-        """
-
-        small_grid_symbol_logo = base64read("small-grid-symbol-logo.png")
-
-        return f"""
-        <style>
-            {fonts_css}
-
-            .syft-container {{
-                padding: 5px;
-                font-family: 'Open Sans';
-            }}
-            .syft-alert-info {{
-                color: #1F567A;
-                background-color: #C2DEF0;
-                border-radius: 4px;
-                padding: 5px;
-                padding: 13px 10px
-            }}
-            .syft-code-block {{
-                background-color: #f7f7f7;
-                border: 1px solid #cfcfcf;
-                padding: 0px 2px;
-            }}
-            .syft-space {{
-                margin-top: 1em;
-            }}
-        </style>
-        <div class="syft-client syft-container">
-            <img src="{small_grid_symbol_logo}" alt="Logo"
-            style="width:48px;height:48px;padding:3px;">
-            <h2>Welcome to {self.name}</h2>
-            <div class="syft-space">
-                <!-- <strong>Institution:</strong> TODO<br /> -->
-                <!-- <strong>Owner:</strong> TODO<br /> -->
-                <strong>URL:</strong> {getattr(self.connection, 'url', '')}<br />
-                <!-- <strong>PyGrid Admin:</strong> TODO<br /> -->
-            </div>
-            <div class='syft-alert-info syft-space'>
-                &#9432;&nbsp;
-                This domain is run by the library PySyft to learn more about how it works visit
-                <a href="https://github.com/OpenMined/PySyft">github.com/OpenMined/PySyft</a>.
-            </div>
-            <h4>Commands to Get Started</h4>
-            {command_list}
-        </div><br />
-        """
 
     def _fetch_node_metadata(self, credentials: SyftSigningKey) -> None:
         metadata = self.connection.get_node_metadata(credentials=credentials)
