+++ packages/syft/src/syft/client/enclave_client.py
@@ -12,8 +12,7 @@ from .client import SyftClient
 
 @serializable()
 class EnclaveClient(SyftClient):
-    def __repr__(self) -> str:
-        return f"<EnclaveClient: {self.name}>"
+    # TODO: add widget repr for enclave client
 
     @property
     def code(self) -> Optional[APIModule]:
