+++ packages/syft/src/syft/util/filterwarnings.py
@@ -26,3 +26,13 @@ warnings.filterwarnings(
     ".*libuv only supports millisecond timer resolution*",
     category=UserWarning,
 )
+
+# site-packages/IPython/core/formatters.py:367: FormatterWarning:
+#  text/html formatter returned invalid type <class 'type'> (expected <class 'str'>)
+#  for object: <EnclaveClient - cranky_norvig <a2fa3fe86007429189dcd88081185bd5>: HTTPConnection: http://localhost:8082>
+# during printing of client object
+warnings.filterwarnings(
+    "ignore",
+    ".*text/html formatter returned invalid type <class 'type'>*",
+    category=UserWarning,
+)
